requirements.txt
streamlit
pandas
scikit-learn
nltk
joblib
streamlit
tensorflow-macos
numpy

